# pkg/query/flamegraph_arrow: Trim children if below trim threshold