To get rid of entries in the flame graph that are smaller than a pixel (and thus can't be rendered) the code iterates over all rows and checks a row's children cumulative values against its own cumulative value. If the difference is so drastic we remove that child.

Not yet implemented is rewriting the arrow record after trimming. Given the little overhead we can always improve the rendering by trimming too small children. Based on some heuristic (like if more than 1% is trimmed) we may consider recreating the arrow record with just the necessary rows. This would mostly be a bandwidth optimization.

The differences with the added code won't be noticeable for now.
```
name      old time/op    new time/op    delta
Query-24    24.7ms ± 4%    25.7ms ± 8%    ~     (p=0.421 n=5+5)

name      old alloc/op   new alloc/op   delta
Query-24    14.1MB ± 0%    14.2MB ± 0%  +0.30%  (p=0.032 n=5+5)

name      old allocs/op  new allocs/op  delta
Query-24      119k ± 0%      120k ± 0%    ~     (p=0.095 n=5+5)
```