@metalmatze 2023-08-30
Superseded by https://github.com/parca-dev/parca/pull/3722